completing full architecture

+++ idgl/utils/generic_utils.py
@@ -68,12 +68,7 @@ def batch_normalize_adj(mx, mask=None):
     """Row-normalize matrix: symmetric normalized Laplacian"""
     # mx: shape: [batch_size, N, N]
 
-    # strategy 1)
-    # rowsum = mx.sum(1)
-    # r_inv_sqrt = torch.pow(rowsum, -0.5)
-    # r_inv_sqrt[torch.isinf(r_inv_sqrt)] = 0. # I got this error: copy_if failed to synchronize: device-side assert triggered
-
-    # strategy 2)
+    # we obtain the norm of each column vector ||a_i||
     rowsum = torch.clamp(mx.sum(1), min=VERY_SMALL_NUMBER)
     r_inv_sqrt = torch.pow(rowsum, -0.5)
     if mask is not None:
@@ -83,8 +78,10 @@ def batch_normalize_adj(mx, mask=None):
     for i in range(r_inv_sqrt.size(0)):
         r_mat_inv_sqrt.append(torch.diag(r_inv_sqrt[i]))
 
+    mx = mx.float()
     r_mat_inv_sqrt = torch.stack(r_mat_inv_sqrt, 0)
     return torch.matmul(torch.matmul(mx, r_mat_inv_sqrt).transpose(-1, -2), r_mat_inv_sqrt)
+
 
 def normalize_sparse_adj(mx):
     """Row-normalize sparse matrix: symmetric normalized Laplacian"""
